chore: move prompt to legacy
