+++ examples/embeddings/knowledge_base/main.go
@@ -12,9 +12,9 @@ import (
 	"github.com/henomis/lingoose/legacy/chat"
 
 	"github.com/henomis/lingoose/index/vectordb/jsondb"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/llm/openai"
 	"github.com/henomis/lingoose/loader"
-	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/textsplitter"
 	"github.com/henomis/lingoose/types"
 )
