+++ examples/embeddings/qdrant/main.go
@@ -8,9 +8,9 @@ import (
 	"github.com/henomis/lingoose/index"
 	indexoption "github.com/henomis/lingoose/index/option"
 	qdrantdb "github.com/henomis/lingoose/index/vectordb/qdrant"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/llm/openai"
 	"github.com/henomis/lingoose/loader"
-	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/textsplitter"
 )
