+++ examples/llm/openai/stream/chat/main.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 
 	"github.com/henomis/lingoose/legacy/chat"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/llm/openai"
-	"github.com/henomis/lingoose/prompt"
 )
 
 func main() {
