+++ examples/pipeline/callbacks/main.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/henomis/lingoose/legacy/pipeline"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/llm/openai"
-	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
