+++ examples/pipeline/history/main.go
@@ -8,8 +8,8 @@ import (
 	"github.com/henomis/lingoose/history"
 	"github.com/henomis/lingoose/legacy/chat"
 	"github.com/henomis/lingoose/legacy/pipeline"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/llm/openai"
-	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
