+++ examples/pipeline/memory/main.go
@@ -8,8 +8,8 @@ import (
 	"github.com/henomis/lingoose/legacy/decoder"
 	"github.com/henomis/lingoose/legacy/memory/ram"
 	"github.com/henomis/lingoose/legacy/pipeline"
+	"github.com/henomis/lingoose/legacy/prompt"
 	llmmock "github.com/henomis/lingoose/llm/mock"
-	"github.com/henomis/lingoose/prompt"
 )
 
 func main() {
