+++ examples/pipeline/openai/main.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/henomis/lingoose/legacy/memory/ram"
 	"github.com/henomis/lingoose/legacy/pipeline"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/llm/openai"
-	"github.com/henomis/lingoose/prompt"
 	"github.com/henomis/lingoose/types"
 )
