+++ examples/prompt/chat/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/henomis/lingoose/legacy/chat"
-	"github.com/henomis/lingoose/prompt"
+	"github.com/henomis/lingoose/legacy/prompt"
 )
 
 func main() {
