+++ examples/prompt/hello/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/henomis/lingoose/prompt"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/types"
 )
