+++ legacy/pipeline/qa/qa.go
@@ -9,7 +9,7 @@ import (
 	indexoption "github.com/henomis/lingoose/index/option"
 	"github.com/henomis/lingoose/legacy/chat"
 	"github.com/henomis/lingoose/legacy/pipeline"
-	"github.com/henomis/lingoose/prompt"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/types"
 )
