+++ legacy/pipeline/sql/sql.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/henomis/lingoose/legacy/decoder"
 	"github.com/henomis/lingoose/legacy/pipeline"
-	"github.com/henomis/lingoose/prompt"
+	"github.com/henomis/lingoose/legacy/prompt"
 	"github.com/henomis/lingoose/types"
 )
