+++ legacy/prompt/template.go
