+++ legacy/prompt/whisper.go
